Combat, wheel, party, player and item fixes
Multiple fixes and behavioral adjustments across combat, player, wheel and item code:

- Combat: For Monks casting non-healing spells, apply the weapon's elementalBond only when present (not COMBAT_NONE); pass caster to doCombatHealth in harmonyHeal; compute Monk melee bonus using a double multiplier and round the resulting damage.
- Wheel: Increase unlockedVesselResonances precision by changing from uint8_t to uint16_t (supports 0..10000 = 0..100%); clarify doc comments for Flurry of Blows damage/critical labels.
- Party: When applying guiding presence, add a base offset of 100 to BUFF_MANTRA when setting the shared mantra value.
- Player: clearCooldowns now handles missing instant spells gracefully for group cooldowns (allows clearing group cooldowns even if the spell lookup fails) and guards spell->isSpender()/isBuilder() calls with null checks.
- Item: Adjust displayed speed string to use half the stored speed value; use fieldAbsorbPercent when comparing absorb values across combat fields.

These changes correct gameplay logic, improve numeric precision and prevent null-accesses and misleading UI values.

+++ src/creatures/players/components/wheel/wheel_definitions.hpp
@@ -226,8 +226,8 @@ struct PlayerWheelMethodsBonusData {
 		int damage = 0;
 		int healing = 0;
 	};
-	// value * 100. Example: 1% == 100
-	std::array<uint8_t, 4> unlockedVesselResonances = {};
+	// value * 100. Example: 1% == 100; agora uint16_t (0..10000 para 0..100%)
+	std::array<uint16_t, 4> unlockedVesselResonances = {};
 
 	// Raw value. Example: 1 == 1
 	struct Skills {
